Split session implementation and tests
Move session configuration and append implementation into focused
submodules so `session/mod.rs` stays below the repository line-count
limit while preserving the existing public API.

Split the oversized session test module into focused test files with a
shared fixture module. Keep the existing assertions and behaviours
unchanged while making each test file small enough for the project
style rule.
