+++ chutoro-core/src/session/tests.rs
@@ -6,548 +6,13 @@
 //! They sit at the base of the testing hierarchy and cover the `append` API
 //! at the method level — happy paths, validation failures, partial-progress
 //! semantics, and property-based invariants via [`proptest`].
+//!
+//! [`ClusteringSession`]: super::ClusteringSession
 
-use std::{num::NonZeroUsize, sync::Arc};
-
-use proptest::prelude::*;
-use rstest::{fixture, rstest};
-
-use crate::{
-    CandidateEdge, ChutoroBuilder, ChutoroError, ClusteringSession, CpuHnsw, DataSource,
-    DataSourceError, DataSourceErrorCode, ExecutionStrategy, HnswParams, MetricDescriptor,
-    SessionConfig, SessionRefreshPolicy,
-};
-
-#[derive(Clone, Debug)]
-struct SessionTestSource {
-    values: Vec<f32>,
-    name: &'static str,
-}
-
-impl SessionTestSource {
-    fn with_len(len: usize) -> Self {
-        Self {
-            values: (0..len).map(|value| value as f32).collect(),
-            name: "session-test",
-        }
-    }
-}
-
-impl DataSource for SessionTestSource {
-    fn len(&self) -> usize {
-        self.values.len()
-    }
-
-    fn name(&self) -> &str {
-        self.name
-    }
-
-    fn distance(&self, i: usize, j: usize) -> Result<f32, DataSourceError> {
-        let left = self
-            .values
-            .get(i)
-            .ok_or(DataSourceError::OutOfBounds { index: i })?;
-        let right = self
-            .values
-            .get(j)
-            .ok_or(DataSourceError::OutOfBounds { index: j })?;
-        Ok((left - right).abs())
-    }
-
-    fn metric_descriptor(&self) -> MetricDescriptor {
-        MetricDescriptor::new("session-test:abs")
-    }
-}
-
-#[fixture]
-fn session_builder() -> ChutoroBuilder {
-    ChutoroBuilder::new()
-}
-
-#[test]
-fn builder_defaults_include_session_configuration() {
-    let builder = ChutoroBuilder::new();
-
-    assert_eq!(builder.min_cluster_size(), 5);
-    assert_eq!(builder.hnsw_params(), &HnswParams::default());
-    assert_eq!(
-        builder.session_refresh_policy(),
-        &SessionRefreshPolicy::manual()
-    );
-}
-
-#[test]
-fn refresh_policy_threshold_can_be_set_and_cleared() {
-    let threshold = NonZeroUsize::new(10).expect("threshold must be non-zero");
-    let policy = SessionRefreshPolicy::manual().with_refresh_every_n(Some(threshold));
-
-    assert_eq!(policy.refresh_every_n(), Some(threshold));
-
-    let cleared = policy.with_refresh_every_n(None);
-    assert_eq!(cleared.refresh_every_n(), None);
-}
-
-#[rstest]
-#[case(
-    HnswParams::new(8, 32).expect("params").with_rng_seed(17),
-    SessionRefreshPolicy::manual()
-)]
-#[case(
-    HnswParams::new(12, 48).expect("params").with_rng_seed(91),
-    SessionRefreshPolicy::manual().with_refresh_every_n(NonZeroUsize::new(7))
-)]
-fn build_session_derives_config_from_builder(
-    session_builder: ChutoroBuilder,
-    #[case] hnsw_params: HnswParams,
-    #[case] refresh_policy: SessionRefreshPolicy,
-) {
-    let source = Arc::new(SessionTestSource::with_len(4));
-    let session = session_builder
-        .with_min_cluster_size(9)
-        .with_hnsw_params(hnsw_params.clone())
-        .with_session_refresh_policy(refresh_policy)
-        .build_session(source)
-        .expect("session must build");
-
-    assert_eq!(session.config().min_cluster_size().get(), 9);
-    assert_eq!(session.config().hnsw_params(), &hnsw_params);
-    assert_eq!(session.config().refresh_policy(), &refresh_policy);
-}
-
-#[rstest]
-#[case(0)]
-#[case(3)]
-#[case(6)]
-fn build_session_does_not_validate_current_source_len(
-    session_builder: ChutoroBuilder,
-    #[case] len: usize,
-) {
-    let source = Arc::new(SessionTestSource::with_len(len));
-    let session = session_builder
-        .with_min_cluster_size(5)
-        .build_session(source)
-        .expect("session creation must not validate current source length");
-
-    assert_eq!(session.point_count(), 0);
-    assert_eq!(session.snapshot_version(), 0);
-}
-
-#[rstest]
-fn build_session_rejects_zero_min_cluster_size(session_builder: ChutoroBuilder) {
-    let source = Arc::new(SessionTestSource::with_len(0));
-    let err = session_builder
-        .with_min_cluster_size(0)
-        .build_session(source)
-        .expect_err("zero min_cluster_size must fail");
-
-    assert_eq!(err, ChutoroError::InvalidMinClusterSize { got: 0 });
-}
-
-#[rstest]
-fn build_session_rejects_gpu_preferred_execution_strategy(session_builder: ChutoroBuilder) {
-    let source = Arc::new(SessionTestSource::with_len(2));
-    let err = session_builder
-        .with_execution_strategy(ExecutionStrategy::GpuPreferred)
-        .build_session(source)
-        .expect_err("session creation must remain CPU-only");
-
-    assert_eq!(
-        err,
-        ChutoroError::BackendUnavailable {
-            requested: ExecutionStrategy::GpuPreferred,
-        }
-    );
-}
-
-#[test]
-fn build_session_maps_hnsw_construction_failure_to_cpu_hnsw_failure() {
-    let source = Arc::new(SessionTestSource::with_len(0));
-    let config = SessionConfig::new(
-        NonZeroUsize::new(5).expect("minimum cluster size must be non-zero"),
-        HnswParams::default(),
-        SessionRefreshPolicy::manual(),
-    );
-
-    let err = ClusteringSession::new_failing_for_test(config, source)
-        .expect_err("CpuHnsw construction failure must surface as CpuHnswFailure");
-    assert!(
-        matches!(err, ChutoroError::CpuHnswFailure { .. }),
-        "expected CpuHnswFailure, got {err:?}"
-    );
-}
-
-fn make_session(
-    builder: ChutoroBuilder,
-    source_len: usize,
-) -> (ClusteringSession<SessionTestSource>, Arc<SessionTestSource>) {
-    let source = Arc::new(SessionTestSource::with_len(source_len));
-    let session = builder
-        .build_session(Arc::clone(&source))
-        .expect("session must build");
-    (session, source)
-}
-
-#[rstest]
-fn append_empty_slice_is_noop(session_builder: ChutoroBuilder) {
-    let (mut session, _) = make_session(session_builder, 4);
-
-    session.append(&[]).expect("empty append must succeed");
-
-    assert_eq!(session.point_count(), 0);
-    assert_eq!(session.snapshot_version(), 0);
-    assert!(session.pending_edges.is_empty());
-}
-
-#[rstest]
-fn append_single_index_increases_point_count(session_builder: ChutoroBuilder) {
-    let (mut session, _) = make_session(session_builder, 4);
-
-    session.append(&[0]).expect("first append must succeed");
-
-    assert_eq!(session.point_count(), 1);
-    assert_eq!(session.snapshot_version(), 0);
-    assert!(session.pending_edges.is_empty());
-}
-
-#[rstest]
-fn append_batch_accumulates_direct_harvested_edges(session_builder: ChutoroBuilder) {
-    let hnsw_params = HnswParams::new(4, 16)
-        .expect("HNSW params must be valid")
-        .with_rng_seed(41);
-    let source = Arc::new(SessionTestSource::with_len(6));
-    let mut session = session_builder
-        .with_hnsw_params(hnsw_params.clone())
-        .build_session(Arc::clone(&source))
-        .expect("session must build");
-    let direct_index =
-        CpuHnsw::with_capacity(hnsw_params, source.len()).expect("direct index must allocate");
-    let mut expected_edges = Vec::new();
-    let indices = [0, 1, 2, 3];
-
-    for index in indices {
-        let edges = direct_index
-            .insert_harvesting(index, source.as_ref())
-            .expect("direct insert must succeed");
-        expected_edges.extend(edges);
-    }
-
-    session.append(&indices).expect("batch append must succeed");
-
-    assert_eq!(session.point_count(), indices.len());
-    assert_eq!(session.snapshot_version(), 0);
-    assert_eq!(session.pending_edges, expected_edges);
-}
-
-#[rstest]
-fn append_delegates_duplicate_rejection_to_hnsw(session_builder: ChutoroBuilder) {
-    // Duplicate detection is delegated to CpuHnsw::insert_harvesting, which
-    // returns HnswError::DuplicateIndex, mapped here to ChutoroError::CpuHnswFailure.
-    let (mut session, _) = make_session(session_builder, 2);
-
-    session.append(&[0]).expect("first append must succeed");
-    let err = session
-        .append(&[0])
-        .expect_err("duplicate append must fail");
-
-    assert!(
-        matches!(err, ChutoroError::CpuHnswFailure { .. }),
-        "expected CpuHnswFailure, got {err:?}"
-    );
-    assert_eq!(session.point_count(), 1);
-}
-
-#[rstest]
-fn append_rejects_out_of_bounds_index(session_builder: ChutoroBuilder) {
-    let (mut session, source) = make_session(session_builder, 2);
-
-    let err = session
-        .append(&[source.len()])
-        .expect_err("out-of-bounds append must fail");
-
-    assert!(
-        matches!(err, ChutoroError::DataSource { .. }),
-        "expected DataSource error, got {err:?}"
-    );
-    assert_eq!(
-        err.data_source_code(),
-        Some(DataSourceErrorCode::OutOfBounds)
-    );
-    assert_eq!(session.point_count(), 0);
-}
-
-#[rstest]
-fn append_failure_preserves_prior_successes(session_builder: ChutoroBuilder) {
-    // Use ≥3 source points so that inserting index 1 into a graph that
-    // already contains index 0 produces at least one harvested edge.
-    // We mirror append_batch_accumulates_direct_harvested_edges to establish
-    // the expected edge set independently via CpuHnsw::insert_harvesting.
-    let hnsw_params = HnswParams::new(4, 16)
-        .expect("HNSW params must be valid")
-        .with_rng_seed(41);
-    let source = Arc::new(SessionTestSource::with_len(4));
-    let mut session = session_builder
-        .with_hnsw_params(hnsw_params.clone())
-        .build_session(Arc::clone(&source))
-        .expect("session must build");
-
-    // Build the expected edge set using the direct index as a baseline.
-    let direct_index =
-        CpuHnsw::with_capacity(hnsw_params, source.len()).expect("direct index must allocate");
-    let mut expected_edges = Vec::new();
-    for index in [0usize, 1] {
-        let edges = direct_index
-            .insert_harvesting(index, source.as_ref())
-            .expect("direct insert must succeed");
-        expected_edges.extend(edges);
-    }
-    // expected_edges is non-empty at this point because inserting index 1
-    // into a graph that already contains index 0 harvests their mutual edge.
-    assert!(
-        !expected_edges.is_empty(),
-        "test precondition: at least one edge must be harvested from the first two insertions"
-    );
-
-    // Now append the two valid indices followed by one out-of-bounds index.
-    let err = session
-        .append(&[0, 1, source.len()])
-        .expect_err("out-of-bounds index must cause failure");
-
-    assert!(
-        matches!(err, ChutoroError::DataSource { .. }),
-        "expected DataSource error, got {err:?}"
-    );
-    assert_eq!(
-        err.data_source_code(),
-        Some(DataSourceErrorCode::OutOfBounds)
-    );
-    assert_eq!(
-        session.point_count(),
-        2,
-        "the two successful insertions must be preserved"
-    );
-    assert_eq!(session.snapshot_version(), 0);
-    assert_eq!(
-        session.pending_edges, expected_edges,
-        "harvested edges from successful insertions must survive a later failure"
-    );
-}
-
-#[rstest]
-fn append_does_not_publish_label_snapshot(session_builder: ChutoroBuilder) {
-    let (mut session, _) = make_session(session_builder, 4);
-
-    session.append(&[0, 1, 2]).expect("append must succeed");
-
-    assert_eq!(session.snapshot_version(), 0);
-}
-
+mod append;
+mod builder;
+mod common;
+mod concurrency;
 #[cfg(feature = "metrics")]
-#[rstest]
-fn append_records_deterministic_latency_via_clock_seam(session_builder: ChutoroBuilder) {
-    use std::{sync::Arc, time::Duration};
-
-    use metrics_util::debugging::{DebugValue, DebuggingRecorder};
-
-    // Install a local debugging recorder for this test only.
-    let recorder = DebuggingRecorder::new();
-    let snapshotter = recorder.snapshotter();
-    metrics::with_local_recorder(&recorder, || {
-        // Build a session with a fixed clock that reports exactly 5 ms per point.
-        let (session, _) = make_session(session_builder, 4);
-        let clock = Arc::new(crate::session::clock::FixedMonotonicClock::with_elapsed(
-            Duration::from_millis(5),
-        ));
-        let mut session = session.with_clock_for_test(clock);
-
-        session.append(&[0]).expect("single append must succeed");
-    });
-
-    let snapshot = snapshotter.snapshot();
-    let histogram = snapshot
-        .into_hashmap()
-        .into_iter()
-        .find(|(key, _)| key.key().name() == "chutoro.session.append.point_seconds")
-        .map(|(_, (_, _, value))| value)
-        .expect("histogram must be recorded");
-
-    if let DebugValue::Histogram(buckets) = histogram {
-        let recorded = buckets
-            .first()
-            .expect("at least one sample must be recorded")
-            .into_inner();
-        assert!(
-            (recorded - 0.005_f64).abs() < 1e-6,
-            "recorded latency must match the fixed clock: expected 0.005 s, got {recorded}"
-        );
-    } else {
-        panic!("expected a Histogram metric value, got {histogram:?}");
-    }
-}
-
-#[rstest]
-fn concurrent_readers_observe_consistent_point_count(session_builder: ChutoroBuilder) {
-    // Append two points on the writer thread, then share the session
-    // read-only across multiple concurrent threads.
-    let (mut session, _) = make_session(session_builder, 4);
-    session.append(&[0, 1]).expect("append must succeed");
-    let shared = Arc::new(std::sync::RwLock::new(session));
-
-    let handles: Vec<_> = (0..8)
-        .map(|_| {
-            let shared = Arc::clone(&shared);
-            std::thread::spawn(move || {
-                let guard = shared.read().expect("read lock must not be poisoned");
-                (guard.point_count(), guard.snapshot_version())
-            })
-        })
-        .collect();
-
-    for handle in handles {
-        let (point_count, snapshot_version) = handle.join().expect("reader thread must not panic");
-        assert_eq!(
-            point_count, 2,
-            "all readers must observe the same point_count"
-        );
-        assert_eq!(snapshot_version, 0, "snapshot_version must remain 0");
-    }
-}
-
-#[rstest]
-fn snapshot_version_is_immutable_under_concurrent_readers(session_builder: ChutoroBuilder) {
-    // snapshot_version must read as 0 under all concurrent readers;
-    // append does not publish a label snapshot.
-    let (mut session, _) = make_session(session_builder, 4);
-    session.append(&[0, 1, 2]).expect("append must succeed");
-    let shared = Arc::new(std::sync::RwLock::new(session));
-
-    let handles: Vec<_> = (0..16)
-        .map(|_| {
-            let shared = Arc::clone(&shared);
-            std::thread::spawn(move || {
-                shared
-                    .read()
-                    .expect("read lock must not be poisoned")
-                    .snapshot_version()
-            })
-        })
-        .collect();
-
-    for handle in handles {
-        let version = handle.join().expect("reader thread must not panic");
-        assert_eq!(
-            version, 0,
-            "snapshot_version must be immutable across all readers"
-        );
-    }
-}
-
-proptest! {
-    /// Any non-zero `min_cluster_size` in `[1, 1000]` must yield a session whose
-    /// `config().min_cluster_size()` equals the requested value.
-    #[test]
-    fn build_session_preserves_min_cluster_size(size in 1usize..=1000) {
-        let source = Arc::new(SessionTestSource::with_len(0));
-        let session = ChutoroBuilder::new()
-            .with_min_cluster_size(size)
-            .build_session(source)
-            .expect("must build for any non-zero min_cluster_size");
-        prop_assert_eq!(session.config().min_cluster_size().get(), size);
-    }
-
-    /// A freshly-built session must always start with `point_count() == 0` and
-    /// `snapshot_version() == 0`, regardless of source length.
-    #[test]
-    fn build_session_always_starts_empty(len in 0usize..=256) {
-        let source = Arc::new(SessionTestSource::with_len(len));
-        let session = ChutoroBuilder::new()
-            .build_session(source)
-            .expect("must build for any source length");
-        prop_assert_eq!(session.point_count(), 0);
-        prop_assert_eq!(session.snapshot_version(), 0);
-    }
-
-    /// Zero `min_cluster_size` must always return `InvalidMinClusterSize`,
-    /// regardless of any other builder state.
-    #[test]
-    fn build_session_always_rejects_zero_min_cluster_size(len in 0usize..=256) {
-        let source = Arc::new(SessionTestSource::with_len(len));
-        let err = ChutoroBuilder::new()
-            .with_min_cluster_size(0)
-            .build_session(source)
-            .expect_err("zero must always fail");
-        prop_assert!(
-            matches!(err, ChutoroError::InvalidMinClusterSize { got: 0 }),
-            "expected InvalidMinClusterSize {{ got: 0 }}, got {err:?}"
-        );
-    }
-
-    /// Appending a generated unique sequence of in-bounds indices must make
-    /// `point_count()` match the number of appended points.
-    #[test]
-    fn append_unique_indices_updates_point_count(
-        indices in proptest::collection::vec(0usize..32, 0..=32)
-            .prop_map(|mut indices| {
-                indices.sort_unstable();
-                indices.dedup();
-                indices
-            })
-    ) {
-        let source = Arc::new(SessionTestSource::with_len(32));
-        let mut session = ChutoroBuilder::new()
-            .with_hnsw_params(HnswParams::default().with_rng_seed(73))
-            .build_session(source)
-            .expect("session must build");
-
-        session
-            .append(&indices)
-            .expect("unique in-bounds append sequence must succeed");
-
-        prop_assert_eq!(session.point_count(), indices.len());
-        prop_assert_eq!(session.snapshot_version(), 0);
-    }
-
-    /// For any unique, in-bounds index sequence of length ≥ 1, `append`
-    /// must buffer exactly the candidate edges that direct
-    /// `CpuHnsw::insert_harvesting` produces for the same sequence.
-    #[test]
-    fn append_pending_edges_match_direct_harvested_edges(
-        indices in proptest::collection::vec(0usize..16, 1..=16)
-            .prop_map(|v| {
-                let mut seen = std::collections::HashSet::new();
-                v.into_iter()
-                    .filter(|index| seen.insert(*index))
-                    .collect::<Vec<_>>()
-            })
-    ) {
-        let source = Arc::new(SessionTestSource::with_len(16));
-        let hnsw_params = HnswParams::new(4, 16)
-            .expect("HNSW params must be valid")
-            .with_rng_seed(99);
-
-        // Build expected edge set via the direct index.
-        let direct_index = CpuHnsw::with_capacity(hnsw_params.clone(), source.len())
-            .expect("direct index must allocate");
-        let mut expected_edges: Vec<CandidateEdge> = Vec::new();
-        for &index in &indices {
-            let edges = direct_index
-                .insert_harvesting(index, source.as_ref())
-                .expect("direct insert must succeed");
-            expected_edges.extend(edges);
-        }
-
-        // Append the same sequence via the session.
-        let mut session = ChutoroBuilder::new()
-            .with_hnsw_params(hnsw_params)
-            .build_session(Arc::clone(&source))
-            .expect("session must build");
-        session
-            .append(&indices)
-            .expect("unique in-bounds append sequence must succeed");
-
-        prop_assert_eq!(
-            session.pending_edges,
-            expected_edges,
-            "pending_edges must match direct harvested edges for the same index sequence"
-        );
-    }
-}
+mod metrics;
+mod properties;
